add new "Simple Dialogue" sample
