+++ Editor/obj.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fc0698cee00156148b3843afd0c5dec5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
