+++ Samples~/Simple Dialogue.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4591dfefe8f11074696154a73216165a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
