+++ Samples~/Simple Dialogue/Dialogue.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e094463a8bde7fb4ba7e2e5ccd12c43e
+guid: 1492920a1c9e68c4f9dc7fd921400853
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
