+++ Samples~/Simple Dialogue/Dialogue/dial_simple.ibra
@@ -12,4 +12,4 @@ Oh...
 [Ava]
 I know!
 [Ava]
-Why don't we go <color=blue>fishing</color>?
+Why don't we go <color=yellow>fishing</color>?
