+++ Samples~/Simple Dialogue/Dialogue/dial_simple.ibra.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c73a8071bc1840c48833c2a801557459
+guid: f910a7d5b681d91449e33c59a20888ad
 ScriptedImporter:
   internalIDToNameTable: []
   externalObjects: {}
