+++ Samples~/Simple Dialogue/Dialogue/dial_string_functions.ibra
