+++ Samples~/Simple Dialogue/Dialogue/dial_string_functions.ibra.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 990558897ffcc3e48b6412b2402c24dc
+guid: 1250cfa8764ba104bb8eb4e9685ae4a5
 ScriptedImporter:
   internalIDToNameTable: []
   externalObjects: {}
