+++ Samples~/Simple Dialogue/ExampleScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c4e3d292cd201ca4c8dca45d53b9b7ec
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
