+++ Samples~/Simple Dialogue/Prefabs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 246ca61282bfe6f4abdeb15346f714be
+guid: 5544d65dc820fb04dbe56c36b1ace5f4
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
