+++ Samples~/Simple Dialogue/Prefabs/ChoiceButton.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fcf20762e0dcf3e43ab5735e387a239d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
