+++ Samples~/Typewriter Dialogue/Assets.meta
