+++ Samples~/Typewriter Dialogue/Assets/Fonts.meta
