+++ Samples~/Typewriter Dialogue/Assets/Fonts/alegreya-sans.asset.meta
