+++ Samples~/Typewriter Dialogue/Assets/Fonts/alegreya-sans.extrabold.ttf.meta
