+++ Samples~/Typewriter Dialogue/Assets/Resources.meta
