+++ Samples~/Typewriter Dialogue/Assets/Resources/Sprites.meta
