+++ Samples~/Typewriter Dialogue/Assets/Resources/Sprites/happy.png
