+++ Samples~/Typewriter Dialogue/Assets/Resources/Sprites/worried.png
