+++ Samples~/Typewriter Dialogue/Assets/Sprites.meta
