+++ Samples~/Typewriter Dialogue/Assets/Sprites/DialogueBoxSimple.png.meta
