+++ Samples~/Typewriter Dialogue/Dialogue.meta
