+++ Samples~/Typewriter Dialogue/Dialogue/dial_function_example.ibra
@@ -1,3 +1,4 @@
+# this example only works with the TypewriterDialogueManager!
 [NPC]
 Psssst kid. New feature dropped.
 [NPC]
