+++ Samples~/Typewriter Dialogue/Dialogue/dial_function_example.ibra.meta
