+++ Samples~/Typewriter Dialogue/Dialogue/dial_test_variable.ibra
