+++ Samples~/Typewriter Dialogue/ExampleScene.unity.meta
