+++ Samples~/Typewriter Dialogue/Scripts.meta
