+++ Samples~/Typewriter Dialogue/Scripts/ExampleDialogueInput.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 
 namespace Ibralogue.Examples
 {
-    public class DialogueInput : MonoBehaviour
+    public class ExampleDialogueInput : MonoBehaviour
     {
-        [SerializeField] private DefaultDialogueManager dialogueManager;
+        [SerializeField] private SimpleDialogueManager dialogueManager;
         private void Update()
         {
             if (Input.GetKeyDown(KeyCode.Return) || Input.GetKeyDown(KeyCode.KeypadEnter))
